+++ srcs/channel/channel.hpp
@@ -1,12 +1,26 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   channel.hpp                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: crisfern <crisfern@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/09/28 11:22:31 by crisfern          #+#    #+#             */
+/*   Updated: 2023/09/28 11:22:33 by crisfern         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef CHANNEL_HPP
 #define CHANNEL_HPP
 #define MAX_NUMBER_OF_CHN 100	
 #include <string>
 #include <iostream>
 #include "../server/server.hpp"
+#include "../command/reply.hpp"
 #include <algorithm>
 
 class	user;
+class 	server;
 class	channel {
 
 	private:
@@ -16,12 +30,13 @@ class	channel {
 		std::vector<user> 		list_of_members;
 		std::string 			_password;
 		std::string 			_mode;
-		int 					_user_limit;
+		size_t 					_user_limit;
 
 		channel 			( void );
 	public:
 
 		channel 			( std::string str );
+		channel 			( std::string c_name, std::string password );
 		channel 			( const channel & var );
 		~channel			( void );
 		channel &operator=	(const channel &tmp);
@@ -33,7 +48,8 @@ class	channel {
 		std::vector<user> get_list_of_members(void) const {return(this->list_of_members);};
 		std::string get_password(void) const {return (this->_password);};
 		std::string get_mode(void) const {return (this->_mode);};
-		int get_user_limit(void) const {return (this->_user_limit);};
+		size_t get_user_limit(void) const {return (this->_user_limit);};
+		user   *get_user_from_nick(std::string name);
 
 		/*###########################################
 		#				SETTERS						#
@@ -41,7 +57,7 @@ class	channel {
 		void set_topic(std::string topic) {this->_topic = topic;};
 		void set_password(std::string pass) {this->_password = pass;};
 		void set_mode(std::string mode) {this->_mode = mode;};
-		void set_user_limit(int i) {this->_user_limit = i;};
+		void set_user_limit(size_t i) {this->_user_limit = i;};
 
 		/*###########################################
 		#				FUNCTIONS					#
@@ -49,6 +65,9 @@ class	channel {
 		void	add_member(user &usr);
 		void	rmv_member(user &usr);
 		bool	is_user_in_channel(const user &usr);
+		bool 	is_user_operator(const user &usr);
+		void	set_user_operator(const user &usr, const bool &flag);
+		void 	parse_mode_flag(user &usr, std::string &modes, std::vector<std::string> mode_params, server &srv);
 };
 std::ostream &operator<<(std::ostream& os, const channel &tmp);
