+++ srcs/command/command.hpp
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   command.hpp                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: crisfern <crisfern@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/09/28 11:13:47 by crisfern          #+#    #+#             */
+/*   Updated: 2023/09/28 11:13:56 by crisfern         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef COMMAND_HPP
 #define COMMAND_HPP
 #include <string>
@@ -18,5 +30,10 @@ namespace cmd
 	void topic(server &svr, int poll_fd_pos, std::string str);
 	void invite(server &svr, int poll_fd_pos, std::string str);
 	void kick(server &svr, int poll_fd_pos, std::string str);
+	void notice(server &svr, int poll_fd_pos, std::string str);
+	void part(server &svr, int poll_fd_pos, std::string str);
+	void mode(server &svr, int poll_fd_pos, std::string str);
+	void names(server &svr, int poll_fd_pos, std::string str);
+	void list(server &svr, int poll_fd_pos, std::string str);
 }
 #endif
