+++ srcs/command/nick.cpp
@@ -1,5 +1,28 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   nick.cpp                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: crisfern <crisfern@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/09/28 11:16:27 by crisfern          #+#    #+#             */
+/*   Updated: 2023/09/28 11:16:36 by crisfern         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "command.hpp"
 #include "reply.hpp"
+
+// Parameters: <nickname> [ <hopcount> ]
+//
+//   is used to give user a nickname or change the previous one.
+//
+//   The <hopcount> parameter must be ignored.
+//
+//      Numeric Replies:
+//
+//           ERR_NONICKNAMEGIVEN             ERR_ERRONEUSNICKNAME
+//           ERR_NICKNAMEINUSE
 
 bool invalid_nick(std::string str)
 {
@@ -14,21 +37,21 @@ void cmd::nick(server &svr, int poll_fd_pos, std::string str) {
   user &usr = svr.get_user(pollfd.fd);
   //set nick
   user *nick;
-  if (str == "NICK")
+  if (str == "")
   {
-	  svr.send_message(": 431: No nickname was given \r\n", usr.get_fd());
+	  svr.send_message(ERR_NONICKNAMEGIVEN, usr.get_fd());
 	  return ;
   }
   if (invalid_nick(str))
   {
-	  svr.send_message(": 432 " + str + " : Erroneus nickname \r\n", usr.get_fd());
+	  svr.send_message(ERR_ERRONEUSNICKNAME(str), usr.get_fd());
 	  return ;
   }
   if ((nick = svr.get_user_from_nick(str)) != 0)
   {
     if (nick->get_nick().compare(usr.get_nick()) == 0)
       return ;
-	  svr.send_message(": 433 " + str + " : Nickname is already in use \r\n", usr.get_fd());
+	  svr.send_message(ERR_NICKNAMEINUSE(str), usr.get_fd());
 	  return ;
   }
   usr.set_nick(str);
