+++ srcs/command/quit.cpp
@@ -1,12 +1,22 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   quit.cpp                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: crisfern <crisfern@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/09/28 11:19:08 by crisfern          #+#    #+#             */
+/*   Updated: 2023/09/28 11:19:17 by crisfern         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "command.hpp"
 #include "reply.hpp"
+#include "../server/server.hpp"
 
 void  cmd::quit(server &svr, int poll_fd_pos, std::string str) {
   poll_fd pollfd = svr.get_pollfd(poll_fd_pos);
   user &usr = svr.get_user(pollfd.fd);
-  // BUG si no pasamos parámetros str == QUIT
-  // Qué pasa si quiero que mi mensaje sea QUIT?
-  // Este comportamiento se repite para cada comando
   if (str.empty())
 	str += "Cierra al salir (Mensaje por defecto del server)";
   str += "\r\n";
