--- tester/command/command.cpp
@@ -1,10 +1,0 @@
-#include "command.hpp"
-#include "reply.hpp"
-
-// NICKNAME
-
-
-
-
-
-
